+++ BP/scripts/config/fluids.js
@@ -6,18 +6,12 @@ const fluidItems = {
         type: "nether_star_essence",
         output: "minecraft:bucket",
     },
-    "utilitycraft:saline_coolant_bucket": {
-        amount: 1000,
-        type: "saline_coolant",
-        output: "minecraft:bucket",
-    },
 };
 
 const fluidHolders = {
     "minecraft:bucket": {
         types: {
             nether_star_essence: "utilitycraft:nether_star_essence_bucket",
-            saline_coolant: "utilitycraft:saline_coolant_bucket",
         },
         required: 1000,
     },